Fix compile warning
Fixed the following compile warning in commander:
    Compile [  2.7%]: commander.cpp
    [Warning] commander.cpp@111,0: comparison between signed and
                   unsigned integer expressions [-Wsign-compare]

+++ commander.cpp
@@ -108,7 +108,7 @@ void Commander::input_handler()
     //if the user hit enter
     if (nkey == 13) {
         //walk the callbacks and call them one at a time
-        for (int n = 0; n<cmd._vready.size();n++) {
+        for (size_t n = 0; n<cmd._vready.size();n++) {
             cmd._vready[n]();
         }
     }
